Make CellProfiler interop semantics nominal
Move the CellProfiler benchmark bridge away from scattered heuristics and stringly payload handling toward load-bearing runtime contracts. Core runtime semantics now carry explicit object-label domains, slice/plane alignment, measurement lookup dialects, object/label feature roles, Zernike descriptor semantics, and artifact-payload validation strategies. Object-label validation is now keyed by ArtifactKind and accepts ObjectLabelPayload/ObjectLabelSet nominally instead of treating them as incidental array-like values, which fixes slice-aligned object-label outputs without weakening generic array/table validation.

Extend the CellProfiler adapter and converter around those contracts. Module execution now records richer declared object domains, measurement subjects, source-image context, runtime-plane projection, and generated module semantics. Converter settings for structuring elements, watershed, expand/shrink, image math, and module defaults are centralized into nominal settings/semantic abstractions so generated functions reuse OpenHCS contracts instead of one-off dictionaries or hardcoded assumptions.

Add core-backed CellProfiler processing semantics. Smoothing backend selection is registered by SmoothingMethod, radial intensity distribution measurement uses a native CellProfiler-compatible backend strategy, object-intensity-distribution delegates geometry to the backend family, and Zernike descriptor equivalence is represented as an explicit object-measurement stability contract. Exact intensity Zernikes delegate to the Centrosome-compatible backend while debug traces are typed and env-gated.

Broaden runtime equivalence and artifact behavior. Measurement feature parsing now recognizes indexed object Zernike descriptors, object-label location projections preserve declared plane ID domains, runtime payload normalization can aggregate slice-aligned object labels into plane-scoped payloads, spatial grids carry nominal topology, source matching uses stable source identity, and measurement lookup goes through explicit dialects.

Benchmark progress: reran the official 22-case set plus cp_tutorial_3d_monolayer with cached native references at /tmp/openhcs_labmeeting_prelim/official22_plus_monolayer_validation_after_nominal_payload_20260508_current. Current parity is 9/23: ExampleHuman, the four illumination examples, ExampleImagingFlowCytometryObjectsInGrid, ExamplePercentPositive, ExampleTumor, and ExampleUntangleAndStraightenWorms. The earlier execution blockers for ExampleSpeckles and both UntangleWorms cases are gone; those cases now reach semantic comparison and fail on measurement differences. Execution-speed target passes 19/23, while total-phase speed target passes 6/23; Grid is parity-clean but remains below the 4x execution-speed target at about 3.60x.

Validation: python compile gate over 119 changed/untracked Python files passed; git diff --check passed excluding the local LibreOffice lock file; focused runtime equivalence tests for declared plane object domains and Zernike/shape descriptors passed; focused runtime-values payload tests passed; focused UntangleWorms module execution test passed; focused three-case blocker benchmark completed with no native/openhcs execution errors; full 23-case benchmark completed with observations=23 and no execution errors.

+++ benchmark/cellprofiler_library/functions/convertobjectstoimage.py
@@ -107,7 +107,7 @@ class Uint16ImageModeRenderer(ImageModeRenderer):
         colormap_value: str,
     ) -> np.ndarray:
         del colormap_value
-        return labels.astype(np.float32)
+        return labels.astype(np.int32, copy=False)
 
 
 def _get_colormap(colormap_name: str, num_labels: int) -> np.ndarray:
